mvp_h2/requirements.txt
pandas
numpy
scikit-learn
seaborn
plotly
fastapi
uvicorn
pyyaml
